requirements.txt
azure-identity
openai
python-dotenv
pydantic
rich
beautifulsoup4
pymupdf4llm
azure-ai-inference
